Add check and ability to provide betas and model fit.
We decided to allow users to provide their own betas along with a
model fit object, but this creates some difficulties. These changes
need to be further tested, because I am a little uncertain we covered
all edge cases.

In particular, I think we might want to revisit the logic of the
central if-else statements. I think the assumption for the top-level
else that the model fit is provided is not longer true.