Sets open_progress to false.
This hardcodes open_progress = FALSE into the calls to
rstan::sampling. This is because some recent changes to rstan and/or R
was trying to open a browser on systems where that was not supported,
and thus failing the tests. We might consider removing this or making
it an argument to the sampling functions.