Remove now unnecessary Stan models.
The two "omni" models (better name needed) include code to handle both
random and fixed effects in the same model. By doing this, we are
halving the number of files and models that we need to maintain.

This was done in preparation of allowing:

  1) user defined priors
  2) random slopes and intercepts

Now, these changes can be made in a single place, rather than having
to pass the changes into multiple files.