Refactor the random effects to allow for random slopes and interepts.
This commit changes how we are getting the matrix of random effects,
as well as the grouping indices for the hyperparameters. Before, we
were pulling them from the result of lFormula, specifically from the
reTrms element. This worked for the random intercepts only.

This is roughly modeled after the code in the rstanarm package, but
uses mapply() to make for more efficient code.

We will need to add more tests to ensure the random slopes work as
expected, but they should.