Fix two serious bugs in the Stan model for random effects.
There were two typos which were contributing to the random effects to
be calculated incorrectly.

First, there was an issue in the function make_random_betas where we
were indexing by n rather than i. This was making all the random
effects the same number, effectively replacing the intercept and
causing issues with the model fit.

Second, in the transformed parameters, we were applying this function
to the raw betas, rather than the ones which had been centered. This
likewise made the random effects swap with the intercept term.

We need to figure out a better way of testing the Stan models to
ensure we are getting the correct answers.