Completely alter how the intercept and priors are handled.
In order to start using a different prior distribution for the
predictors, we needed to change how the intercept is handled. This is
because if we wanted to put a shrinkage prior on the betas, for
example, we would not want that shrinkage applied to the intercept
term. Therefore, we do not want to include the intercept into the
predictor matrix.

This creates some difficulties, because we now need to check if there
is an intercept, remove it from the predictor matrix (currently not
implemented) and then get the prior on the intercept to be separate
from the betas.

To implement this, we looked to rstanarm for inspiration, as the authors have
already coded a lot of this functionality into their package.