Adds code to center parameters and handle separate intercept.
In order to implement many other changes(e.g. RHS priors), we need to be able to
separate the intercept from the other predictors. However, we need to
do this in a robust way that allows users to omit the intercept in the
model formula, e.g.

eDNA_lm(Cq ~ -1 ... )

One option is to do what rstanarm does, which is to handle the
intercept as a special term. However, this requires many changes to
handle the case where the only predictor provide is the intercept.

The other option, and the one implemented here, is to detect if there
is an intercept term at the beginning of the model matrix, and if so
handle it as a special case.

However, there is one more change needed - we need to figure out how
the prior for the intercept should be handled. Since we are centering
the predictors, the intercept term will be the "centered" intercept,
and hense the prior on that centered term can end up being quite
strange.

Need to think further about this.