Adds code to deal with intercepts separately.
To implement the changes we need for the horseshoe priors, we need to
handle the intercept separately from the rest of the predictors. This
adds the code to the Stan model to handle this. It also centers the
predictors.

The corresponding R code needs to be updated to handle these changes.