Follows rstanarm and scales by both x and y.
This follows the priors scaling as explained in ?rstanarm::priors.

However, I am not sure that scaling makes sense in this case, as the
response, y, is CQ, but the model parameters are on the latent
parameter, log eDNA. We might want to circle back to this to find
better defaults.