Adds varying measurement error.
This adds varying measurement error, which allows for the measurement
error on CQ to increase as CQ approaches the threshold. It uses a
linear model to do this where sigma_Cq = alpha + beta * centered_Cq.

The centered_Cq value is centered to 30, which appears to be
approximately an "average" value - CQ rarely goes below 25 in these
data. We might want to adjust this in the future.

Other changes which are needed:

- Handle the varying measurement error in predict functions
- Make the computations more efficient