Make changes to allow for multiple standard curves.
This allows multiple standard curves to be provided. This is
particularly helpful when we are trying to analyze data from multiple
experiments, or when the lab was forced to use different standards
midway through running the qPCR.

When a single set of parameters is provided, they are rep()'ed to the
length of the response variables.

We might need to add in a check for when the parameters are not = 1,
and also not = N.