Remove test which is no longer relevant.
With the change to allow multiple standard curves, we needed to change
the behavior of est_p_detect. In particular, when a model_fit was
provided and there were multiple curves, the user needed to provide a
single one for calculation of p(detect).

This new expected behavior is covered in test_std_curve.R.