Order the data so the model can be vectorized.
Before, the model loops over the data one obs. at a time. However, we
can vectorize the model by arranging the data into two blocks, at the
threshold or below it. Then, the data can be segmented inside the
model.

More optimizations might be possible to segment the data in the
transformed data block, rather than inside the model block.