Separate prep functions for model and sim.
We will need to clean this up. The functions for simulation and for
modeling have diverged to the point that having the same prep function
handle both cases was betting to be an issue. The prep_data function
was getting too complex, and it was creating too much friction to make
non-breaking changes as the model evolved.

We might need to revisit this as we update the simulation functions,
and eventually might merge these back together. But for now, it is
slowing down development.