Change p_detect to be a user-provided value rather than parameter.
We added the ability for artemis to model the data using a
zero-inflated model, where there was a separate process which lead to
zero detections.

Unfortunately, most eDNA studies result in a lot of non-detections,
and the most parsimonious model fit would be a large probability of
zero values from this process, rather than by the model
parameters. Unless there were a large number of detections, it became
difficult to estimate p_zero. This manifest as a large percent of
divergent transitions.

As a result, we have moved the p_zero to be a user-specified
constant. This still allows the zero-inflated model, but prevents the
model from blowing up.

When there are lots of data, we can provide a means to estimate this
parameter later.