Update model structure to ensure varying meas. error works.
The code to allow the measurement error to vary is a basic linear
relationship centered at Cq = 30, e.g.

sigma = sigma + adjustment * (Cq - 30)

This works well in most cases, but can lead to degenerate cases where
the result is negative. This led to the iteration being rejected by
Stan's sampler. In rstan and older versions of cmdstan, the value for
this iteration in the results would be NaN. This broke most summary
and plotting functions.

Rather than just add na.rm to the relevant functions, we decided to
change the model to make it impossible for the sigma value to be
negative. We did this by making the model on sigma

sigma = exp(sigma_raw + adjustment * (Cq - 30))