Adding back features from the old model.
Continuing to add back features which were present in the old
model. To simplify the code, we have separated eDNA_lm and
eDNA_lm_zinf - this makes it clear to the user when the model is a
zero-inflated model. It also simplifies the code in each model,
although there is now repeated code in places. However, the previous
model had no repeated code, but was difficult to understand and debug.

In this case, a little repetition is worth the increase in readability.