Change sigma_CQ to sigma_ln_eDNA.
This change was needed because when the model was formulated with the
measurement error on Cq, we ran into some strange behavior when
multiple standard curves were used.

Specifically, with multiple standard curves, the Cq -> ln(eDNA)
conversion would have differences between observations. This means
that even though all observations had the same value of sigma_Cq in
the model, they effectively were slightly different.

The solution was to term the residual error from ln(eDNA). Since the
obs are converted to this from the various std curve parameters, the
measurement error would not change between obs. This is more
consistent with standard regression models.

However, we still need to fully update the documentation to reflect
these changes.