Fixes #32.
This makes the predict function automatically look for the appropriate
columns when "newdata" is provided. Additionally, it will drop the
random effects from model formula.

We should still probably issues some warning when a random effect
model is provided that we do not currently support "redraw" of new
random effects in predictions. Long term, we might add this, but it is
low priority at the moment.