Move install info to README to avoid strange install vignette.
With the install guide a vignette, it will show up in the list of
vignettes inside R after artemis is already installed, which is
strange. Likewise, it shows up on the CRAN webpage for the package in
the first position, which is likewise strange.

By moving it to the README, it will be visible to people visiting the
github repo, where it will be needed most.