Revert some changes to the model.
This commit reverts the changes to the random effects. We played
around with using the QR decomp on the random effects, however this is
not quite the proper means to estimate these effects. For a detailed
explanation for how these effects could be more robustly estimates,
see https://cran.r-project.org/web/packages/lme4/vignettes/Theory.pdf.

Instead, we implement the non-centered parameterization for the random
effects to increase the MCMC efficiency. This is has been better in
testing, but could still be improved.

Lastly, we experiment with a boundary-avoiding prior on the variance
of the random effects. This probably needs more testing and refinement.