Adds sparse matrix calc to the random effects model.
This commit refactors how random effects are handled in the Stan code
to use a compressed row storage mode. This is preferred in cases when
there are large, sparse matrices. This change speeds up model fitting
when there are large numbers of random effects, but likely slows down
the model when there are few.

As eDNA experiments tend to collect many Filters, and the Filters are
often specified as random effects, this should speed up model fits
overall.

In brief testing, it resulted in a 4x slowdown in models with few
random effects, but a 20x speedup in models with many (500+).

In the future, we could create some conditional statement in the model
to check for the size of the random effects matrix, and only apply
this when needed. We also could add an argument to the modeling
functions to specify if the user wants to use the compressed row storage.