Updates functions to ensure models will be compiled.
This might be too agressive. It avoids the issues brought about when
the models are not compiled correctly, or are compiled but in the
wrong place.

That said, it also causes files to be compiled and written to the
cache directory without user interaction, which is not always polite.

Maybe revisit this later.