Update methods and classes to fully use cmdstanr.
This is the final work required to pull out the remaining rstan
dependencies and switch over to using cmdstanr. The biggest change
here is the methods handling the fit model object.

There is remaining work to be done on the plotting, but for now this
should work for fitting models.