Fix issues with summary functions - need to add back functionality.
This changes the summary.eDNA model to utilize the cmdstanr summary()
function. This is better because it adds a lot of functionality and
also removes some clunky code from artemis.

We still want to display the variable names on the beta values, so
that functionality will need to be added back in, likely by just
updating the value in the "variable" column to the names from the
predictor matrix.