Adds full zip component to random effect.
This adds the full zero-inflated model to the version with random
effects. At the moment, we only allow a fixed-effects model on the zip
component - we might expand this to allow for full random effects
eventually.

There are some issues still:

1. We need to test this code for accuracy - testing random effects can
be quite challenging, so we need to be careful here.

2. There are some issues with how random effects are being handled in
the functions.