Adds a Poisson (count) model to artemis.
We have added a Poisson model to artemis. This addition was
necessitated by the fact that some qPCR assays are sensitive to very
low concentrations (average single counts of DNA in solution). At
these low levels, the dynamics of sampling do not fit the assumed
normal distribution. Instead, a Poisson distribution is more appropriate.

Under the assumption of a Poisson distribution, the possible copy
numbers sampled from a solution with average 1 copy will include
0,1,2, and 3 copies. Under qPCR dynamics, these can result in
drastically different Cq/CT values.

An additional benefit of using the Poisson model when copy numbers are
low is that it captures the occurance of 0 copies inherently rather
than requiring a separate process to capture a potential zero-inflated
process.