Adds code to call est_p_detect with count models.
This updates the est_p_detect to handle count models. This is a
breaking change, as we need to differentiate between the count and
regular lm version of the est_p_detect functions.

It might make sense to change the name of est_p_detect back, however
this could create ambiguity. We'll need to think a bit about what the
best move forward.